+++ src/models/UserModels/userResponse.ts
@@ -2,6 +2,6 @@ export interface User {
   id: string;
   username: string;
   role: string;
-  last_login: string;
-  login_count: string;
+  last_login?: string;
+  login_count?: string;
 }