+++ src/service/userService.ts
@@ -35,4 +35,10 @@ export default {
     accessToken,
     queryParams: { page: page.toString() }
   }) as Promise<User[]>,
+
+  resetAllQuests: async (accessToken: string) => await sendRequest({
+    method: 'DELETE',
+    endpoint: `${basePath}/completed-quests`,
+    accessToken
+  }),
 }